Move config classes to config package instead of one-class-per-package
