+++ molgenis-platform-integration-tests/src/test/java/org/molgenis/integrationtest/platform/PlatformITConfig.java
@@ -25,8 +25,8 @@ import org.molgenis.data.postgresql.identifier.EntityTypeRegistryPopulator;
 import org.molgenis.data.security.SystemEntityTypeRegistryImpl;
 import org.molgenis.data.transaction.TransactionManager;
 import org.molgenis.data.validation.ExpressionValidator;
+import org.molgenis.integrationtest.config.ScriptTestConfig;
 import org.molgenis.integrationtest.data.TestAppSettings;
-import org.molgenis.integrationtest.script.ScriptTestConfig;
 import org.molgenis.jobs.JobConfig;
 import org.molgenis.jobs.JobExecutionConfig;
 import org.molgenis.jobs.JobFactoryRegistrar;
